OE-5383: Add validation for firm/global rights.

+++ protected/models/User.php
@@ -591,8 +591,17 @@ class User extends BaseActiveRecordVersioned
         }
     }
 
+    /**
+     * @param array $firms
+     * @throws CDbException
+     * @throws FirmSaveException
+     */
     public function saveFirms(array $firms)
     {
+        if(!$this->global_firm_rights && count($firms) === 0){
+            throw new FirmSaveException('When global firm rights are not set, a firm must be selected');
+        }
+
         $transaction = Yii::app()->db->beginTransaction();
         FirmUserAssignment::model()->deleteAll('user_id = :user_id', array('user_id' => $this->id));
         foreach($firms as $firm){
